LayoutTests/imported/w3c/web-platform-tests/streams/readable-streams/cancel.sharedworker-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

FAIL cancel.js shared worker wrapper file ReferenceError: Can't find variable: SharedWorker
